+++ lib/get_navigation/src/routes/not_found_page.dart
@@ -1,0 +1,40 @@
+import 'package:flutter/material.dart';
+
+class NotFoundPage extends StatelessWidget {
+  const NotFoundPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: [
+            Text(
+              '404',
+              style: TextStyle(
+                fontSize: 48,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+            SizedBox(width: 16),
+            Container(
+              height: 48,
+              width: 1,
+              color: Colors.grey,
+            ),
+            SizedBox(width: 16),
+            Text(
+              'This page could not be found.',
+              style: TextStyle(
+                fontSize: 16,
+                color: Colors.black54,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}