+++ lib/get_rx/get_rx.dart
@@ -4,6 +4,5 @@
 /// defining reactive types, and working with reactive workers.
 library;
 
-export "src/rx_stream/rx_stream.dart";
 export "src/rx_types/rx_types.dart";
 export "src/rx_workers/rx_workers.dart";