+++ lib/get_rx/src/rx_stream/rx_stream.dart
@@ -1,8 +1,0 @@
-library;
-
-import 'dart:async';
-
-import '../rx_typedefs/rx_typedefs.dart';
-
-//part 'get_stream.dart';
-part 'mini_stream.dart';